% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-vote.R
\name{combine_enrichment_results}
\alias{combine_enrichment_results}
\title{Combine Enrichment Results}
\usage{
combine_enrichment_results(enrich_results, databases, verbose = FALSE)
}
\arguments{
\item{enrich_results}{List of enrichment results from run_enrichment.}

\item{databases}{Character vector of databases.}

\item{verbose}{Logical, whether to print progress messages.}
}
\value{
A list of combined result tables for each database.
}
\description{
Combines enrichment results from multiple gene lists using harmonic mean p-value.
}
