% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTM-class.R
\name{create_eQTM}
\alias{create_eQTM}
\title{Create an eQTM object}
\usage{
create_eQTM(data, metadata = list())
}
\arguments{
\item{data}{A data.frame containing eQTM data with columns: cpg, statistics, p_value, distance, and at least one of entrez or ensembl.}

\item{metadata}{A list of metadata (optional).}
}
\value{
An eQTM object.
}
\description{
Create an eQTM object
}
