% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTM-class.R
\docType{class}
\name{eQTM-class}
\alias{eQTM-class}
\title{eQTM Class}
\description{
A class to store eQTM data for pathway analysis. eQTM stands for Expression Quantitative Trait Methylation.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame containing eQTM data with columns: cpg, statistics, p_value, distance, and at least one of entrez or ensembl.}

\item{\code{metadata}}{A list of metadata (e.g., data source, time point). Reserved for future use.}
}}

