% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-vote.R
\name{gene_filter}
\alias{gene_filter}
\title{Filter eQTM Data by Statistics and Distance}
\usage{
gene_filter(eQTM, stat_threshold, distance)
}
\arguments{
\item{eQTM}{An eQTM object.}

\item{stat_threshold}{Numeric, the minimum threshold for the statistics value.}

\item{distance}{Numeric, the maximum distance threshold.}
}
\value{
A list with filtered entrez IDs and p-values.
}
\description{
Filters eQTM data based on statistics and distance thresholds.
}
