% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTM-class.R
\name{getData}
\alias{getData}
\alias{getData,eQTM-method}
\title{Get expression quantitative trait methylation (eQTM) Data}
\usage{
getData(object)

\S4method{getData}{eQTM}(object)
}
\arguments{
\item{object}{An eQTM object.}
}
\value{
A data.frame stored in the object.
}
\description{
Retrieve the eQTM data.frame from an eQTM object.
}
