% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTM-class.R
\name{getMetadata}
\alias{getMetadata}
\title{Get metadata}
\usage{
getMetadata(object)
}
\arguments{
\item{object}{An eQTM object.}
}
\value{
The metadata list stored in the eQTM object.
}
\description{
Get metadata
}
