% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTM-class.R
\name{getMetadata}
\alias{getMetadata}
\alias{getMetadata,eQTM-method}
\title{Get expression quantitative trait methylation (eQTM) Metadata}
\usage{
getMetadata(object)

\S4method{getMetadata}{eQTM}(object)
}
\arguments{
\item{object}{An eQTM object.}
}
\value{
A list containing metadata.
}
\description{
Retrieve the metadata list from an eQTM object.
}
